backend/requirements.txt
fastapi==0.105.0
uvicorn==0.23.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
python-multipart==0.0.6
pyyaml==6.0.1
pillow==10.1.0
python-jose==3.3.0
passlib==1.7.4